requirements.txt
customtkinter>=5.2.2
PyMuPDF>=1.26.0
Pillow>=10.0.0
reportlab>=4.0.0